% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{combine_queries}
\alias{combine_queries}
\title{Combine Queries}
\usage{
combine_queries(..., operator = "must")
}
\arguments{
\item{...}{Query lists to combine}

\item{operator}{Boolean operator: "must" (AND), "should" (OR), or "must_not" (NOT)}
}
\value{
A combined query list
}
\description{
Combine multiple queries using boolean logic
}
\examples{
# Female humpback whales from 2020-2023
species <- filter_by_species("Megaptera novaeangliae")
sex <- filter_by_sex("female")
years <- filter_by_year_range(2020, 2023)

query <- combine_queries(species, sex, years, operator = "must")
}
