% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTOmillisecond.R
\name{dateTOmillisecond}
\alias{dateTOmillisecond}
\title{Transform a vector of date to a vector of millisecond.}
\usage{
dateTOmillisecond(date, origin = "1970-01-01", format = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{A character vector represent a period of time from date[1] to date[2].}

\item{origin}{A point of time which set to be zero in millisecond.}

\item{format}{A format for date and origin arguments.}
}
\description{
This function is to transform a period of time in date to millisecond according to the origin date and the format of date.
}
\details{
A vector of size two represent a period of time. The start date of the period will be transform to the first millisecond of the date and the end of the period to the last millisecond.
}
\examples{
dateTOmillisecond(date=c("1970-01-01","1970-02-01"))
#When the first date is after the second one
dateTOmillisecond(date=c("1970-02-01","1970-01-01"))

#examples in different formats
dateTOmillisecond(date=c("1970-01-01","1970-02-01"),format="\%Y-\%d-\%m")
dateTOmillisecond(date=c("1970/01/01","1970/02/01"),format="\%Y/\%m/\%d")
}

