% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{field_exists}
\alias{field_exists}
\title{Field Exists}
\usage{
field_exists(field)
}
\arguments{
\item{field}{Field name}
}
\value{
A query list
}
\description{
Create a query to find documents where a field exists and has a value
}
\examples{
# Find encounters with an assigned individual
query <- field_exists("individualId")
}
