% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{field_missing}
\alias{field_missing}
\title{Field Missing}
\usage{
field_missing(field)
}
\arguments{
\item{field}{Field name}
}
\value{
A query list
}
\description{
Create a query to find documents where a field is missing or null
}
\examples{
# Find encounters without an assigned individual
query <- field_missing("individualId")
}
