% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{filter_by_individual}
\alias{filter_by_individual}
\title{Filter by Individual}
\usage{
filter_by_individual(individual_id)
}
\arguments{
\item{individual_id}{Individual UUID}
}
\value{
A query list
}
\description{
Create a query to find all encounters for a specific individual
}
\examples{
query <- filter_by_individual("123e4567-e89b-12d3-a456-426614174000")
}
