% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{filter_by_sex}
\alias{filter_by_sex}
\title{Filter by Sex}
\usage{
filter_by_sex(sex)
}
\arguments{
\item{sex}{Sex value (e.g., "male", "female", "unknown")}
}
\value{
A query list
}
\description{
Create a query to filter by sex
}
\examples{
query <- filter_by_sex("female")
}
