% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{filter_by_submitter}
\alias{filter_by_submitter}
\title{Filter by Submitter}
\usage{
filter_by_submitter(submitter_id)
}
\arguments{
\item{submitter_id}{User UUID}
}
\value{
A query list
}
\description{
Create a query to find encounters submitted by a specific user
}
\examples{
query <- filter_by_submitter("user-uuid-here")
}
