% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{filter_by_year_range}
\alias{filter_by_year_range}
\title{Filter by Year Range}
\usage{
filter_by_year_range(start_year = NULL, end_year = NULL)
}
\arguments{
\item{start_year}{Minimum year (inclusive), optional}

\item{end_year}{Maximum year (inclusive), optional}
}
\value{
A query list
}
\description{
Create a query to filter by year range
}
\examples{
# Encounters from 2020 to 2023
query <- filter_by_year_range(2020, 2023)

# Encounters from 2020 onwards
query <- filter_by_year_range(start_year = 2020)
}
