% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterstring.R
\name{filterstring}
\alias{filterstring}
\title{Generate the JDOQL part of a filter.}
\usage{
filterstring(filtername, filtervalues, logic = "||", bridge = "==")
}
\arguments{
\item{filtername}{A character which is for the responding variable name in Wildbook framework.}

\item{filtervalues}{A vector of the value for the filter.}

\item{logic}{A parameter which can be "&&" for the logical AND or "||" for the logical OR.}

\item{bridge}{An operator to connect the name and the default value is "==".}
}
\description{
This function is to generate the string in the JDOQL query for some given value of a filter.
}

