% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Identify operating system as Mac, *nix, or Windows}
\usage{
get_os()
}
\value{
Operating system name (character string).
}
\description{
Copied from Hadley Wickham's rappdir package at his suggestion. The
rappdir package currently does not export this function
}

