% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWildbook-package.r
\docType{package}
\name{RWildbook}
\alias{RWildbook}
\alias{RWildbook-package}
\title{RWildbook}
\description{
The primary objective of this package is to provide an R interface with the Wildbook mark-recapture ecological database framework. It helps users to pull data from the Wildbook framework and format data for further analysis with mark-recapture applications like Program MARK(which can be accessed via the RMark package in R).
}

