% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexstring.R
\name{sexstring}
\alias{sexstring}
\title{Generate the JDOQL part for the sex filter.}
\usage{
sexstring(sex)
}
\arguments{
\item{sex}{The value for the sex filter.}
}
\description{
This function is to generate the sex related portion in the JDOQL query.
}
\examples{
sexstring("male")
sexstring(sex="female")
sexstring(sex="unknown")
sexstring(c("male","female"))
sexstring(c("female","male"))
sexstring(c("female","unknown"))
sexstring(c("male","unknown"))
sexstring(sex=c("female","male","unknown"))
}

