% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{text_search}
\alias{text_search}
\title{Text Search}
\usage{
text_search(field, text, fuzzy = FALSE)
}
\arguments{
\item{field}{Field name to search}

\item{text}{Text to search for}

\item{fuzzy}{Whether to use fuzzy matching (default: FALSE)}
}
\value{
A query list
}
\description{
Create a text search query
}
\examples{
# Search for "beach" in locality
query <- text_search("verbatimLocality", "beach")

# Fuzzy search
query <- text_search("verbatimLocality", "beach", fuzzy = TRUE)
}
