% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vignette_1_data}
\alias{vignette_1_data}
\title{Data for the first vignette.}
\format{An object of class \code{data.frame} with 69 rows and 65 columns.}
\source{
{http://whaleshark.org}
}
\usage{
vignette_1_data
}
\description{
Sample data for the first vignette. Originally pulled from 
\url{http://whaleshark.org}, the data has been modified to protect the innocent.
}
\keyword{datasets}

