% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBsearchURL.R
\name{WBsearchURL}
\alias{WBsearchURL}
\title{Generate the search URL given the JDOQL query.}
\usage{
WBsearchURL(username, password, baseURL, jdoql)
}
\arguments{
\item{username}{The username in the Wildbook framework.}

\item{password}{The password in the Wildbook framework.}

\item{baseURL}{The URL represent the desire wildbook data base.}

\item{jdoql}{The JDOQL string for data searching.}
}
\description{
This function helps users to generate the URL for data searching in the Wildbook framework with the account information of Wildbook, the URl of the desire wildbook and the JDOQL query.
}

