% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTOmillisecond.R
\name{dateTOmillisecond}
\alias{dateTOmillisecond}
\title{Transform a vector of date to a vector of millisecond.}
\usage{
dateTOmillisecond(date, origin = "1970-01-01", format = "\%Y-\%m-\%d",
  interval = TRUE)
}
\arguments{
\item{date}{A character vector represent a period of time from date[1] to date[2].}

\item{origin}{A point of time which set to be zero in millisecond.}

\item{format}{A format for date and origin arguments.}

\item{interval}{A logical argument which equals to TRUE by default when the function transfer an period of dates to millisecond.
When interval=FALSE, the function returns the last minute of every elements of date to millisecons.}
}
\description{
This function is to transform a period of time in date to millisecond according to the origin date and the format of date.
}
\details{
A vector of size two represent a period of time. The start date of the period will be transform to the first millisecond of the date and the end of the period to the last millisecond.
}

