% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchWB.R
\name{searchWB}
\alias{searchWB}
\title{Pull data from the Wildbook framework.}
\usage{
searchWB(searchURL = NULL, username = NULL, password = NULL, baseURL,
  jdoql = NULL, object = "encounter", location = NULL,
  locationID = NULL, sighting_date = c("1964-01-01", "2016-12-31"),
  encounter_submission_dates = c("2003-01-01", "2016-12-31"),
  date_format = "\%Y-\%m-\%d", sex = c("male", "female", "unknown"),
  status = c("alive", "dead"), measurement = NULL, individualID = NULL,
  encounterID = NULL, encounter_type = NULL, Date_of_birth = NULL,
  Date_of_death = NULL)
}
\arguments{
\item{searchURL}{A URL for data searching in the Wildbook framework.}

\item{username}{The username in the Wildbook framework.}

\item{password}{The password in the Wildbook framework.}

\item{baseURL}{The URL represent the wildbook data base.}

\item{jdoql}{The JDOQL string for data searching.}

\item{object}{can be either "encounter" for the encounter search or "individual" for the individual search.}

\item{location}{A string of character contained in location names.}

\item{locationID}{A character vector for filtering the locationID.}

\item{sighting_date}{A character for filtering encounters which are sighted during a period of time.}

\item{encounter_submission_dates}{A character for filtering encounters which are submitted during a period of time.}

\item{date_format}{The format for all the arguments of date valule.}

\item{sex}{A character vector of maximum size of three represents the value for the sex filter.}

\item{status}{A character vector of maximum size of two represents the value for the encounter status.}

\item{measurement}{A numeric object sets the minimum individual measurement when searching in the Wildbook framework.}

\item{individualID}{A character vector for searching data of specific individual ID.}

\item{encounterID}{A character vector for searching data of specific encounter ID.}

\item{encounter_type}{A character vector of maximum size of three for searching data with specific encounter type.}

\item{Date_of_birth}{A character vector for searching data of individual which is borned during a period of time.}

\item{Date_of_death}{A character vector for searching data of individual which is dead during a period of time.}
}
\description{
This function allows users to pull data from the Wildbook framework in R.
}
\examples{
##same searches for users of different background.
data1 <- searchWB(searchURL = "http://xinxin:changeme@whaleshark.org/api/jdoql?SELECT FROM org.ecocean.MarkedIndividual WHERE individualID == 'A-001'")

data2 <- searchWB(username="xinxin",password="changeme",baseURL ="whaleshark.org",jdoql="SELECT FROM org.ecocean.Encounter WHERE catalogNumber != null &&  (( dateInMilliseconds >= -189388800000 )&&( dateInMilliseconds <= 1483228740000 ))&&(( dwcDateAddedLong >= 1041379200000 )&&( dwcDateAddedLong <= 1483228740000 ))&&(( individualID == 'a-001' )) ")

data3 <- searchWB(username="xinxin",password="changeme",baseURL ="whaleshark.org",object="Encounter",individualID=c("A-001"))

}

