% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexstring.R
\name{sexstring}
\alias{sexstring}
\title{Generate the JDOQL part for the sex filter.}
\usage{
sexstring(sex)
}
\arguments{
\item{sex}{The value for the sex filter.}
}
\description{
This function is to generate the sex related portion in the JDOQL query.
}

